requirements.txt
cryptography
flask
flask-vite
ipython
numpy
plotly
pywebview
zeroconf